Move Slack Reactions from Demo App to Slack Clone
