+++ DemoApp/Screens/AppConfigViewController/AppConfigViewController.swift
@@ -11,8 +11,6 @@ struct DemoAppConfig {
     var isHardDeleteEnabled: Bool
     /// A Boolean value to define if Atlantis will be started to proxy HTTP and WebSocket calls.
     var isAtlantisEnabled: Bool
-    /// A Boolean value to define if reactions should be rendered Slack-like.
-    var isSlackReactionsEnabled: Bool
 }
 
 class AppConfig {
@@ -25,8 +23,7 @@ class AppConfig {
         // Default DemoAppConfig
         demoAppConfig = DemoAppConfig(
             isHardDeleteEnabled: false,
-            isAtlantisEnabled: false,
-            isSlackReactionsEnabled: false
+            isAtlantisEnabled: false
         )
     }
 }
@@ -100,7 +97,6 @@ class AppConfigViewController: UITableViewController {
     enum DemoAppConfigOption: String, CaseIterable {
         case isHardDeleteEnabled
         case isAtlantisEnabled
-        case isSlackReactionsEnabled
     }
 
     enum ChatClientConfigOption: String, CaseIterable {
@@ -198,10 +194,6 @@ class AppConfigViewController: UITableViewController {
             cell.accessoryView = makeSwitchButton(demoAppConfig.isAtlantisEnabled) { [weak self] newValue in
                 self?.demoAppConfig.isAtlantisEnabled = newValue
             }
-        case .isSlackReactionsEnabled:
-            cell.accessoryView = makeSwitchButton(demoAppConfig.isSlackReactionsEnabled) { [weak self] newValue in
-                self?.demoAppConfig.isSlackReactionsEnabled = newValue
-            }
         }
     }
 
@@ -215,8 +207,6 @@ class AppConfigViewController: UITableViewController {
         case .isHardDeleteEnabled:
             break
         case .isAtlantisEnabled:
-            break
-        case .isSlackReactionsEnabled:
             break
         }
     }
