--- DemoApp/StreamChat/Components/SlackReactions/SlackReactionsMessagePopupVC.swift
@@ -1,15 +1,0 @@
-//
-// Copyright © 2022 Stream.io Inc. All rights reserved.
-//
-
-import StreamChat
-import StreamChatUI
-
-final class SlackReactionsMessagePopupVC: ChatMessagePopupVC {
-    override func setUpLayout() {
-        let messageView = messageContentView as! SlackReactionsMessageView
-        messageView.isInPopupView = true
-        
-        super.setUpLayout()
-    }
-}
