+++ Examples/SlackClone/SlackChatMessageContentView.swift
@@ -9,12 +9,38 @@ import UIKit
 final class SlackChatMessageContentView: ChatMessageContentView {
     override var maxContentWidthMultiplier: CGFloat { 1 }
 
+    lazy var slackReactionsView: SlackReactionsView = {
+        let view = SlackReactionsView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+
+    var isInPopupView = false
+
     override func layout(options: ChatMessageLayoutOptions) {
         super.layout(options: options)
 
         mainContainer.alignment = .leading
         bubbleThreadFootnoteContainer.changeOrdering()
         bubbleContentContainer.directionalLayoutMargins = .zero
+
+        if options.contains(.slackReactions) && !isInPopupView {
+            addSubview(slackReactionsView)
+            let bottomPadding: CGFloat = options.contains(.timestamp) ? 0 : 2
+            NSLayoutConstraint.activate([
+                slackReactionsView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 8),
+                slackReactionsView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -8),
+                slackReactionsView.topAnchor.constraint(equalTo: mainContainer.bottomAnchor, constant: bottomPadding),
+                slackReactionsView.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -8),
+                slackReactionsView.widthAnchor.constraint(greaterThanOrEqualToConstant: 40)
+            ])
+        }
+    }
+
+    override func updateContent() {
+        super.updateContent()
+
+        slackReactionsView.content = content
     }
 
     override func createTextView() -> UITextView {
