+++ Examples/SlackClone/SlackReactionsMessagePopupVC.swift
@@ -2,4 +2,14 @@
 // Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
-import Foundation
+import StreamChat
+import StreamChatUI
+
+final class SlackReactionsMessagePopupVC: ChatMessagePopupVC {
+    override func setUpLayout() {
+        let messageView = messageContentView as! SlackChatMessageContentView
+        messageView.isInPopupView = true
+
+        super.setUpLayout()
+    }
+}
